backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
pydantic==2.5.0
PyJWT==2.8.0
pandas==2.1.4
python-dateutil==2.8.2
APScheduler==3.10.4
